btree: explicit (size_t)npages
This cast is safe because:
- we cannot have both npages and npagebytes specified.
- if npages was specified, it's <= SIZE_MAX.
- if npagebytes was specified, it's <= SIZE_MAX, and therefore
	npages = npagebytes / pagelen;
  will also be <= SIZE_MAX since pagelen is a whole number.
- if neither npages nor npagesbytes were specified, the default value is
  of npagebytes is <= SIZE_MAX, and then we're covered by the above point.

+++ kvlds/btree.c
@@ -216,7 +216,7 @@ btree_init(struct wire_requestqueue * Q_lbs, uint64_t npages,
 			goto err1;
 		}
 	}
-	T->poolsz = npages;
+	T->poolsz = (size_t)npages;
 
 	/* Set default key/value lengths if necessary. */
 	if (*keylen == (uint64_t)(-1)) {
